fix : 이체 및 거래 내역 로직 전반적인 구조 개선 및 트랜잭션 처리 추가

+++ src/main/java/server/domain/account/domain/AccountHistory.java
@@ -21,9 +21,9 @@ public class AccountHistory {
 
     private LocalDateTime createdAt;
 
-    private Long amount;
+    private Integer amount;
 
-    private Long remainAmount;
+    private Integer remainAmount;
 
     private String name;
 
@@ -31,3 +31,4 @@ public class AccountHistory {
         SEND, GET
     }
 }
+
